+++ spec/dummy/app/models/wine.rb
@@ -1,7 +1,7 @@
 class Wine < ActiveRecord::Base
   include Fae::Concerns::Models::Base
 
-  has_many :release
+  has_many :releases
   has_many :winemakers
 
   has_many :oregon_winemakers, -> { where(region_type: 1) },