+++ spec/dummy/app/controllers/admin/content_blocks_controller.rb
@@ -3,6 +3,6 @@ class Admin::ContentBlocksController < Fae::StaticPagesController
   private
 
   def fae_pages
-    [HomePage, AboutUsPage]
+    [HomePage, AboutUsPage, ContactUsPage]
   end
 end