+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     + Disabled checkboxes are now untouchable
     + Multiselect headers update available/added items accurately
     + \#57288: Use real links (not anchors) in dummy app
+    + \#57523: Remove errors after image is reuploaded
 
 ## 1.3