+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     + Disabled checkboxes are now untouchable
     + Multiselect headers update available/added items accurately
     + \#57523: Remove errors after image is reuploaded
+    + \#57772: Link admin logo to parent app root path
 
 ## 1.3