+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ## Unreleased
 
 - enhancements
+    + Generated nested table controllers are now empty and inherit from `Fae::NestedBaseController`
 - bugs
     + Disabled checkboxes are now untouchable
     + Multiselect headers update available/added items accurately
     + \#57288: Use real links (not anchors) in dummy app
     + \#57523: Remove errors after image is reuploaded
+    + \#57772: Link admin logo to parent app root path
 
 ## 1.3