+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     + Disabled checkboxes are now untouchable
     + Multiselect headers update available/added items accurately
     + \#57523: Remove errors after image is reuploaded
+    + \#57772: Link admin logo to parent app root path
     + Force `fae_display_field` to be a string when using the front-end search
 
 ## 1.3