+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - enhancements
+    + Reorganize docs for easier access
     + Generated nested table controllers are now empty and inherit from `Fae::NestedBaseController`
     + \#57075: Add a flag to the install generator for internal usage
     + Support multiple languages on static page fields
@@ -10,6 +11,8 @@
     + Disabled checkboxes are now untouchable
     + Multiselect headers update available/added items accurately
     + \#57523: Remove errors after image is reuploaded
+    + \#57772: Link admin logo to parent app root path
+    + Force `fae_display_field` to be a string when using the front-end search
 
 ## 1.3