+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     + Reorganize docs for easier access
     + Generated nested table controllers are now empty and inherit from `Fae::NestedBaseController`
     + \#57075: Add a flag to the install generator for internal usage
+    + Support multiple languages on static page fields
 - bugs
     + Disabled checkboxes are now untouchable
     + Multiselect headers update available/added items accurately