+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     + Generated nested table controllers are now empty and inherit from `Fae::NestedBaseController`
     + \#57075: Add a flag to the install generator for internal usage
     + Support multiple languages on static page fields
+    + \#57288: main nav items will display the paths of the first drop down item by default
 - bugs
     + Disabled checkboxes are now untouchable
     + Multiselect headers update available/added items accurately